pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ait"
version = "2.0.0"
description = "Autonomous Intelligent Trading - Options trading bot with ML and sentiment analysis"
requires-python = ">=3.11,<3.14"
dependencies = [
    # IBKR
    "ib_insync>=0.9.86",
    # Data
    "yfinance>=0.2.36",
    "polygon-api-client>=1.14",
    "pandas>=2.2",
    "numpy>=1.26",
    "pandas-market-calendars>=4.4",
    # ML (lightweight - no TensorFlow/PyTorch)
    "xgboost>=2.0",
    "lightgbm>=4.3",
    "scikit-learn>=1.4",
    # Options Greeks
    "py-vollib>=1.0.1",
    "scipy>=1.12",
    # Sentiment
    "transformers>=4.38",   # FinBERT
    "torch>=2.2",           # CPU-only for FinBERT inference
    "feedparser>=6.0",
    "finnhub-python>=2.4",
    # Notifications
    "aiohttp>=3.9",
    # Dashboard
    "streamlit>=1.31",
    "plotly>=5.19",
    # Config & Logging
    "pydantic>=2.6",
    "pydantic-settings>=2.1",
    "pyyaml>=6.0",
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.2",
    "mypy>=1.8",
]

[project.scripts]
ait = "ait.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
